backend/pyproject.toml
[tool.poetry]
name = "luchoh.com"
version = "0.1.0"
description = "luchoH Photography"
authors = ["Lucho Hristov <luchoh@luchoh.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
sqlalchemy = "^2.0"
pymysql = "^1.0.2"  # Replace mysqlclient with pymysql
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
pydantic = {extras = ["email"], version = "^2.8.2"}
jinja2 = "^3.1.4"
alembic = "^1.13.2"
bcrypt = "^4.1.3"
pydantic-settings = "^2.3.4"
pillow = "^10.4.0"
unidecode = "^1.3.8"

[tool.poetry.dev-dependencies]
pytest = "^7.3.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.2.0"
pylint = "^3.2.4"
flake8 = "^7.1.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"